application scores

+++ database/migrations/2025_09_06_033659_create_scoring_scales_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('criteria_id'); // FK to scholarships (added later)
             $table->string('value');
-            $table->decimal('score', 5, 2);
+            $table->integer('score');
 
             $table->string('created_by')->nullable();
             $table->string('updated_by')->nullable();
